Split tests by controller or responsibility
